pyproject.toml
[tool.poetry]
name = "halftone"
version = "0.1.0"
description = ""
authors = ["capricorn <8250820+capricorn@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.2.0"
numpy = "^1.26.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
